+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 - Initial release
 - Added default Google Books integration
+
+## [0.2.0] - 2026-01-08
+### Added
+- Provider OpenLibrary
+- Search Books and covers for e books (don't require API KEY)